PinotInstanceRestletResource done
